BREAKING table_filters relationship + distance to surname_table
